+++ ui/src/components/admin/bookTypes/BookTypesList.js
@@ -53,11 +53,13 @@ class BookTypesList extends Component {
     }
 
     componentDidMount() {
-        axios.get("http://localhost:9000/api/bookTypes").then(data => {
-            this.setState({
-                bookTypes : data.data
+        if(window.token !== undefined) {
+            axios.get("http://localhost:9000/api/bookTypes").then(data => {
+                this.setState({
+                    bookTypes : data.data
+                });
             });
-        });
+        }
     }
 
      handleClickOpen(id) {
@@ -74,16 +76,18 @@ class BookTypesList extends Component {
     }
 
     handleAgree() {
-        this.setState({
-            open: false
-        });
-
-        axios.delete("http://localhost:9000/api/bookType/" + this.state.deleteId)
-            .then(data => {
-                alert("Usunieto autora!");
-                this.setState({deleteId: null});
-                this.componentDidMount();
+        if(window.token !== undefined) {
+            this.setState({
+                open: false
             });
+
+            axios.delete("http://localhost:9000/api/bookType/" + this.state.deleteId)
+                .then(data => {
+                    alert("Usunieto autora!");
+                    this.setState({deleteId: null});
+                    this.componentDidMount();
+                });
+        }
     }
 
     render() {
